+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This a informal changelog so i can keep track of what im doing.
 - Implemented chapter upgrades (for when a official release upgrades over a scanlator copy automatically)
 - Changed provider score colours a bit
 - Searching & Filtering works now!
+- Allow addition and removal of synonyms / other names (helps with speeding up searching)
+- Add more debugging logs
 
 ## 2026-03-18
 - Fixed up docker builds with gitlab-ci.